Summary: [OPENBUS-2474] [C++] Assistente nao pode aceitar tempo de espera entre tentativas igual a ZERO
+ Aplicacao do patch elaborado pelo tmartins.
 - Reuso de std::invalid_argument.
 - Uso de static_cast.


git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/sdk/cpp/trunk@157110 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ assistant/include/openbus/assistant/detail/shared_state.hpp
@@ -24,7 +24,7 @@
 
 namespace openbus { namespace assistant {
 
-unsigned int const default_retry_wait = 30u;
+float const default_retry_wait = 30.0f;
 
 namespace assistant_detail {
